requirements.txt
flask==3.0.0
pyyaml==6.0.1
aiohttp==3.9.1
pillow==10.1.0
requests==2.31.0
bleak==0.21.1
bleak-retry-connector==3.4.0
home-assistant-bluetooth==1.12.0
